morphology arrange functions
