+++ kornia/morphology/__init__.py
@@ -1,1 +1,3 @@
+from .basic_operators import *
+from .open_close import *
 from .morphology import *
